+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Officially remove the Emscripten backend.
 - On Windows, fix handling of surrogate pairs when dispatching `ReceivedCharacter`.
 - On macOS 10.15, fix freeze upon exiting exclusive fullscreen mode.
+- On iOS, fix null window on initial `HiDpiFactorChanged` event.
 
 # 0.20.0 Alpha 3 (2019-08-14)